libc/misc: Mark lock functions as weak (#1)
The current strong implementation relies on the fact that the linker
won't pick the functions from libc if all of them have already been
implemented by the program. This allows allows a program to override
them without them being strong.

However, this doesn't let another library overwrite them. The need to be
marked as weak functions for that.